requirements.txt
gdown
higher
pandas
pyyaml
scikit-learn
torchvision
tqdm
